pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-threat-hunting-framework"
version = "0.10.0"
description = "Agentic Threat Hunting Framework - Memory and AI for threat hunters"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Sydney Marrone", email = "athf@nebulock.io"}
]
maintainers = [
    {name = "Sydney Marrone", email = "athf@nebulock.io"}
]
keywords = [
    "threat-hunting",
    "security",
    "cybersecurity",
    "mitre-attack",
    "detection-engineering",
    "soc",
    "blue-team",
    "incident-response",
    "threat-intelligence",
    "agentic-ai"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Environment :: Console",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=10.0.0",
    "jinja2>=3.0.0",
    "python-dotenv>=0.19.0",
    "importlib_resources>=5.0.0; python_version < '3.9'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "bandit[toml]>=1.7.0",
    "pre-commit>=3.0.0",
    "types-PyYAML>=6.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]
similarity = [
    "scikit-learn>=1.0.0",
]
splunk = [
    "requests>=2.25.0",
]
# LLM providers (choose one or use litellm for all)
litellm = [
    "litellm>=1.30.0",
]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.30.0",
]
bedrock = [
    "boto3>=1.28.0",
]
ollama = [
    "ollama>=0.3.0",
]
# Convenience: all LLM providers via LiteLLM
llm = [
    "litellm>=1.30.0",
]
# MCP server
mcp = [
    "mcp[cli]>=1.0.0",
]
all = [
    "litellm>=1.30.0",
    "scikit-learn>=1.0.0",
    "requests>=2.25.0",
    "tavily-python>=0.3.0",
    "mcp[cli]>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Nebulock-Inc/agentic-threat-hunting-framework"
Documentation = "https://github.com/Nebulock-Inc/agentic-threat-hunting-framework/tree/main/docs"
Repository = "https://github.com/Nebulock-Inc/agentic-threat-hunting-framework"
Issues = "https://github.com/Nebulock-Inc/agentic-threat-hunting-framework/issues"
Discussions = "https://github.com/Nebulock-Inc/agentic-threat-hunting-framework/discussions"

[project.scripts]
athf = "athf.cli:main"
athf-mcp = "athf.mcp.server:main"

[tool.setuptools]
packages = ["athf", "athf.agents", "athf.agents.llm", "athf.commands", "athf.core", "athf.data", "athf.mcp", "athf.mcp.tools", "athf.utils"]

[tool.setuptools.package-data]
"athf.data" = [
    "templates/*.md",
    "knowledge/*.md",
    "knowledge/**/*.md",
    "prompts/*.md",
    "hunts/*.md",
    "hunts/**/*.md",
    "docs/*.md",
    "docs/**/*.md",
    "integrations/*.md",
    "integrations/**/*.md",
]

[tool.black]
line-length = 127
target-version = ['py38', 'py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 127
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Ignore missing imports for optional dependencies
[[tool.mypy.overrides]]
module = [
    "boto3",
    "boto3.*",
    "tavily",
    "tavily.*",
    "sklearn.*",
    "requests",
    "requests.*",
    "litellm",
    "litellm.*",
    "openai",
    "openai.*",
    "ollama",
    "ollama.*",
    "anthropic",
    "anthropic.*",
    "athf.core.metrics_tracker",
    "mcp",
    "mcp.*",
]
ignore_missing_imports = true
follow_imports = "skip"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, multiple components)",
    "slow: Slow tests (skipped in quick runs)",
]

[tool.coverage.run]
source = ["athf"]
omit = ["*/tests/*", "*/test_*.py"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.bandit]
exclude_dirs = ["tests", "test_*.py", "*/tests/*"]
skips = ["B101"]  # Skip assert_used check (common in tests)
